+++ src/relstorage/adapters/batch.py
@@ -88,7 +88,7 @@ class RowBatcher(object):
                 or self.size_added >= self.size_limit):
             self.flush()
 
-    def select_from(self, columns, table, **kw):
+    def select_from(self, columns, table, suffix='', **kw):
         """
         Handles a query of the ``WHERE col IN (?, ?,)`` type.
 
@@ -105,7 +105,7 @@ class RowBatcher(object):
             filter_subset = filter_values[:self.row_limit]
             del filter_values[:self.row_limit]
             descriptor = [[(table, (filter_column,)), filter_subset]]
-            self._do_batch(command, descriptor, rows_need_flattened=False)
+            self._do_batch(command, descriptor, rows_need_flattened=False, suffix=suffix)
             for row in self.cursor.fetchall():
                 yield row
 
@@ -123,7 +123,7 @@ class RowBatcher(object):
     def _do_deletes(self):
         return self._do_batch('DELETE', sorted(iteritems(self.deletes)))
 
-    def _do_batch(self, command, descriptors, rows_need_flattened=True):
+    def _do_batch(self, command, descriptors, rows_need_flattened=True, suffix=''):
         count = 0
         for (table, columns), rows in descriptors:
             count += len(rows)
@@ -133,7 +133,8 @@ class RowBatcher(object):
             if len(columns) == 1:
                 stmt, params, these_params_need_flattened = self._make_single_column_query(
                     command, table,
-                    columns[0], rows, rows_need_flattened)
+                    columns[0], rows, rows_need_flattened
+                )
             else:
                 row_template = " AND ".join(
                     ("%s=" % (column,)) + self.delete_placeholder
@@ -148,6 +149,9 @@ class RowBatcher(object):
 
             if these_params_need_flattened:
                 params = self._flatten_params(params)
+            stmt += suffix
+            #import threading
+            #print(threading.current_thread(), stmt)
             __traceback_info__ = params
             self.cursor.execute(stmt, params)
         return count