+++ src/relstorage/adapters/postgresql/__init__.py
@@ -22,10 +22,15 @@ assert PostgreSQLAdapter
 assert select_driver
 
 
-def debug_my_locks(cursor): # pragma: no cover
+def debug_my_locks(cursor, exclusive_only=False): # pragma: no cover
     conn = cursor.connection
-    pid = conn.info.backend_pid
-    cursor.execute("""
+    try:
+        pid = conn.info.backend_pid
+    except AttributeError:
+        # connection.info was added to psycopg2 in 2.8.0
+        cursor.execute("SELECT pg_backend_pid()")
+        pid = cursor.fetchall()[0][0]
+    stmt = ("""
     SELECT
          a.datname,
          l.relation::regclass,
@@ -38,10 +43,13 @@ def debug_my_locks(cursor): # pragma: no cover
          l.granted
     FROM pg_stat_activity a
     JOIN pg_locks l ON l.pid = a.pid
-    WHERE a.pid = %s and l.mode like '%%Exclu%%'
-    AND l.locktype <> 'virtualxid'
+    WHERE a.pid = %s
+    AND l.locktype <> 'virtualxid' """
+            + (""" AND l.mode like '%%Exclu%%' """ if exclusive_only else "")
+            + ("""
     ORDER BY a.pid
-    """, (pid,))
+    """))
+    cursor.execute(stmt, (pid,))
     rows = cursor.fetchall()
     rows = '\n'.join('\t'.join((str(i) for i in r)) for r in rows)
     header = '*Locks for %s isomode %s\n' % (pid, conn.isolation_level)